pyproject.toml
[project]
name = "siu3r"
version = "0.1.0"
authors = [{ name = "Qi Xu", email = "insomniaaac@qq.com" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "curope",
    "dacite>=1.9.2",
    "diff-gaussian-rasterization",
    "einops>=0.8.1",
    "gsplat",
    "hydra-colorlog>=1.2.0",
    "hydra-core>=1.3.2",
    "hydra-optuna-sweeper>=1.2.0",
    "imageio[ffmpeg]>=2.37.0",
    "jaxtyping>=0.3.2",
    "kornia>=0.8.1",
    "lightning>=2.5.2",
    "matplotlib>=3.10.3",
    "nerfview>=0.1.3",
    "numpy<2.0.0",
    "open-clip-torch>=2.32.0",
    "opencv-python>=4.11.0.86",
    "panopticapi",
    "plyfile>=1.1.2",
    "pycocotools>=2.0.10",
    "rich>=14.0.0",
    "rootutils>=1.0.7",
    "splines>=0.3.3",
    "torch==2.4.1",
    "torchmetrics>=1.7.3",
    "torchvision==0.19.1",
    "transformers>=4.53.0",
    "viser>=0.2.23",
    "wandb>=0.21.0",
]

[tool.ruff.lint]
ignore = ["E402", "F401", "F841", "E712", "E741", "F722"]

[[tool.uv.index]]
name = "pytorch-cu118"
url = "https://download.pytorch.org/whl/cu118"
explicit = true

[tool.uv.sources]
torch = [
    { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
torchvision = [
    { index = "pytorch-cu118", marker = "sys_platform == 'linux' or sys_platform == 'win32'" },
]
diff-gaussian-rasterization = { git = "https://github.com/rmurai0610/diff-gaussian-rasterization-w-pose.git" }
gsplat = { git = "https://github.com/nerfstudio-project/gsplat.git" }
curope = { path = "src/models/croco/curope" }
panopticapi = { git = "https://github.com/cocodataset/panopticapi.git" }
